removed comments from code, app doesn't have comments functionality but don't have time to work on it anymore. Turning in as is.

+++ models/Post.js
@@ -24,8 +24,8 @@ Post.init(
       defaultValue: DataTypes.NOW,
     },
     votes: {
-      type: DataTypes.FLOAT,
-      defaultValue: 1.00,
+      type: DataTypes.INT,
+      defaultValue: 1,
     },
     user_id: {
       type: DataTypes.INTEGER,
